+++ index.html
@@ -231,7 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+  <script type="module" src="js/main.js"></script>
   <script src="js/function.js"></script>
-  <script src="js/main.js"></script>
 </body>
 </html>
