+++ package.json
@@ -23,5 +23,6 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  }
+  },
+  "type": "module"
 }
